bugfix/ Rename error to error-temp
